add comment function

+++ src/main/java/com/yang/dto/QuestionDto.java
@@ -5,12 +5,12 @@ import lombok.Data;
 
 @Data
 public class QuestionDto {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmt_create;
     private Long gmt_modified;
-    private Integer customer_id;
+    private Long customer_id;
     private Integer comment_volume;
     private Integer read_volume;
     private Integer like_volume;
